Refactor DatasetOwnership service to not hold transient components in singleton scope
