+++ src/infra/core/src/lib.rs
@@ -29,7 +29,6 @@ mod compacting_service_impl;
 mod dataset_changes_service_impl;
 mod dataset_config;
 mod dataset_layout;
-mod dataset_ownership_service_in_memory_state_initializer;
 mod dataset_ownership_service_inmem;
 mod dependency_graph_repository_inmem;
 mod dependency_graph_service_inmem;
@@ -50,7 +49,6 @@ pub use compacting_service_impl::*;
 pub use dataset_changes_service_impl::*;
 pub use dataset_config::*;
 pub use dataset_layout::*;
-pub use dataset_ownership_service_in_memory_state_initializer::*;
 pub use dataset_ownership_service_inmem::*;
 pub use dependency_graph_repository_inmem::*;
 pub use dependency_graph_service_inmem::*;
